+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [6.9.3]
+
+### Changed
+
+- Improve `IGetCitiesParams` sortable fields interface.
+- Improve `IGetCountriesParams` sortable fields interface.
+
+### Deprecated
+
+- Deprecated `IUserNotification` interface. Please user `INotificationUser` instead.
+
 ## [6.9.2]
 
 ### Fixed
@@ -13,7 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Change `IBaseUser` to `IUserBase` interface.
 - Changed `IFeatureOfPackageBase` to `IFeaturePackageBase`.
 
-## [6.10.0]
+## [6.9.1]
 
 ### Fixed